requirements.txt
# Python 3.9-compatible dependencies for this project notebook workflow.

# Core EDA stack
numpy>=1.24,<3
pandas>=2.0,<3
matplotlib>=3.7,<4
opencv-python>=4.8,<5
PyYAML>=6.0,<7

# Notebook/runtime tools
jupyterlab>=4.2,<5
notebook>=7.2,<8
ipykernel>=6.29,<7
nbformat>=5.10,<6
nbclient>=0.10,<1

# Dataset download utility (used by download_dataset.py)
python-dotenv>=1.0,<2
roboflow>=1.1,<2
